LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-unicode-expected.txt


FAIL WebSockets: bufferedAmount for unicode data assert_equals: expected 6 but got 12
